### 2016-07-20 #42 @whedon
Submission: Gillespie.jl: Stochastic Simulation Algorithm in Julia
**Submitting author:** @sdwfrost (<a href="http://orcid.org/0000-0002-5207-9879">Simon Frost</a>)
**Repository:** <a href="http://github.com/sdwfrost/Gillespie.jl" target ="_blank">http://github.com/sdwfrost/Gillespie.jl</a>
**Version:** v0.0.1
