### 2016-07-20 #42 @vchuravy
@sdwfrost I would recommend to register your package to simplify installation and to tag your package.
