### 2016-07-20 #42 @arfon
> @whedon I would offer to review this. How would I proceed?

@vchuravy - great! I've just added you to the JOSS reviewers team. Please do the following before starting this review:

-  Accept the invite to join the 'joss-reviewers' [team](https://github.com/orgs/openjournals/invitation?via_email=1) on GitHub that you should have received.
- Familiarize yourself with the JOSS [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines)
- Complete this request for information about your background: https://github.com/openjournals/joss/issues/138

Any questions, please don't hesitate to get in touch.
