### 2016-07-20 #42 @vchuravy
@sdwfrost I will add my thoughts here:

- [ ] Code-coverage seems not to be working
- [ ] Performance: Could you add some performance numbers comparing the C and R version with the Julia version.
- [ ] Tests: The tests you are using are testing functionality. Would it be possible to add correctness tests?
- [ ] Documentation: It is great to see that you have inline docstrings. Maybe you could use them to generate a documentation (with Documenter.jl)
