### 2016-07-21 #42 @pjotrp
This software consists of 100+ lines of Julia code and is a reimplementation of existing code. I would like to ask the author to confirm here that he thinks this software is a valuable contribution to research. I want to point out that once published this work will be visible for a long time. Adding an example of having used this tool in research would be a good idea.

Also a speed claim should show numbers and a version of 0.01 looks alpha to me.
