### 2016-07-21 3c16552
Fix code coverage
