+++ .travis.yml
@@ -19,4 +19,4 @@ script:
   - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
   - julia -e 'Pkg.clone(pwd()); Pkg.build("Gillespie"); Pkg.test("Gillespie"; coverage=true)'
 after_success:
-  - julia -e 'cd(Pkg.dir("DataFrames")); Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())'
+  - julia -e 'cd(Pkg.dir("Gillespie")); Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())'
