### 2016-07-21 #42 @rveltz
Dear Simon,

I don’t understand the meaning of this mail. Is it some reviews from joss for Gillespie.jl?

Romain

> On 21 Jul 2016, at 12:27, Simon Frost <notifications@github.com> wrote:
>
> @vchuravy <https://github.com/vchuravy> Thanks for the feedback!
>
> The package is on its way to being in METADATA.jl (overlooked version number for Julia in requires)
> Coverage is now fixed (silly error) - at 93% <https://coveralls.io/github/sdwfrost/Gillespie.jl?branch=master>.
> Happy to put in the comparisons. Would R and Rcpp suffice?
> As these are stochastic simulations, it's a bit hard to check. I can test for consistency (such that examples give fixed outputs).
> I haven't used Documenter.jl. Will give it a go.
> Will add DOIs.
> @pjotrp <https://github.com/pjotrp> Thanks for the feedback too!
>
> There aren't many tools out there that allow one to easily run these types of simulations - at least, those that aren't heavily based on SBML. The paper for GillespieSSA only has 28 citations on Google Scholar - not great for a tool - but other packages e.g. hybridModels use it, and other packages (e.g. POMP) have their own custom implementation. The overall style of the package (and Julia-specific ways of speeding things up), was originally borrowed by @rveltz <https://github.com/rveltz>, and we're now working on implementations of more research-level algorithms. So, despite being a small library (it'd be longer in Java ;) ), on balance, I say that it's worth it.
> The version number is following Julia guidelines (as Julia is < 1.0 right now), rather than necessarily reflecting an alpha stage.
> I also see the 'Community' checkbox is unchecked. I'll add a section there too.
>
> Thanks again @vchuravy <https://github.com/vchuravy> @pjotrp <https://github.com/pjotrp> @arfon <https://github.com/arfon> @whedon <https://github.com/whedon> I think that this is a great initiative.
>
