### 2016-07-21 #42 @sdwfrost
Hi @rveltz - I just mentioned you in conversation in my responses to referees (as it says on the bottom of the email) - this also gives you the option to mute the thread.
