### 2016-07-21 #42 @rveltz
Hi,

I did not know you could share reviews like that,

Romain

> On 21 Jul 2016, at 13:34, Simon Frost <notifications@github.com> wrote:
>
> Hi @rveltz <https://github.com/rveltz> - I just mentioned you in conversation in my responses to referees (as it says on the bottom of the email) - this also gives you the option to mute the thread.
>
