### 2016-07-22 b67171e
Updated SSA.jl for benchmark comparisons
