### 2016-07-22 #42 @arfon
Thanks for review @vchuravy and the updates @sdwfrost. @sdwfrost please let me know when you've had a chance to complete the other items.
