+++ docs/make.jl
@@ -1,0 +1,3 @@
+using Documenter, Gillespie
+
+makedocs()
