+++ docs/src/index.md
@@ -1,0 +1,24 @@
+# Gillespie.jl: Stochastic Simulation Algorithm in Julia
+
+```@contents
+```
+
+## Types
+
+```@docs
+SSAStats
+```
+
+```@docs
+SSAArgs
+```
+
+```@docs
+SSAResult
+```
+
+## Functions
+
+```@docs
+ssa
+```
