### 2016-07-22 #42 @vchuravy
@sdwfrost Thank for including the benchmarks. Just for your information `@benchmark` takes the parameters `samples` and `seconds` as limits. So in your case the benchmark is not running for a 1000 samples because the time limit is exceeded, to turn that of you can set `seconds=Inf`. Also it is preferential to write `$x0` instead of `x0`, otherwise you will also benchmark the cost of accessing a global variable (expensive in Julia).

```julia
@benchmark ssa(x0,F,nu,parms,tf) samples=1000
# should be
@benchmark ssa($x0,$F,$nu,$params,$tf) samples=1000 seconds=Inf
```

Also the times for R,Rccp,and SSA should probably be divided by `N` so that they are more readily comparable.
