### 2016-07-22 #42 @sdwfrost
@vchuravy I didn't get round to finishing the writeup on the benchmarks; thanks for the tips on `@benchmark`; I didn't know that you could use interpolation. Working through Documenter.jl right now. Will update here when the rest of the items are complete. Thanks again for your input.
