### 2016-07-22 0acfc5f
More Documenter tweaks
