+++ docs/make.jl
@@ -1,3 +1,9 @@
 using Documenter, Gillespie
 
 makedocs()
+
+deploydocs(
+    deps = Deps.pip("mkdocs","python-markdown-math"),
+    repo = "github.com/sdwfrost/Gillespie.jl.git",
+    julia = "0.4"
+)
