+++ .travis.yml
@@ -20,3 +20,5 @@ script:
   - julia -e 'Pkg.clone(pwd()); Pkg.build("Gillespie"); Pkg.test("Gillespie"; coverage=true)'
 after_success:
   - julia -e 'cd(Pkg.dir("Gillespie")); Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())'
+  - julia -e 'Pkg.add("Documenter")'
+  - julia -e 'cd(Pkg.dir("Gillespie")); include(joinpath("docs", "make.jl"))'
