+++ docs/mkdocs.yml
@@ -1,0 +1,25 @@
+site_name:        Gillespie.jl
+repo_url:         https://github.com/sdwfrost/Gillespie.jl
+site_description: Stochastic simulation algorithm in Julia
+site_author:      sdwfrost
+
+theme: readthedocs
+
+extra_css:
+  - assets/Documenter.css
+
+extra_javascript:
+  - https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML
+  - assets/mathjaxhelper.js
+
+markdown_extensions:
+  - extra
+  - tables
+  - fenced_code
+  - mdx_math
+
+docs_dir: 'build'
+
+pages:
+  - Home: index.md
+
