### 2016-07-22 dd3f870
More fixes
