+++ .travis.yml
@@ -11,12 +11,6 @@ before_script:
   - "export DISPLAY=:99.0"
   - "sh -e /etc/init.d/xvfb start"
   - sleep 3 # give xvfb some time to start
-before_install:
-  - openssl aes-256-cbc -K $encrypted_6c1f315b1f00_key -iv $encrypted_6c1f315b1f00_iv -in .documenter.enc -out .documenter -d
-#matrix:
-#  allow_failures:
-#    - os: osx
-# uncomment the following lines to override the default test script
 script:
   - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
   - julia -e 'Pkg.clone(pwd()); Pkg.build("Gillespie"); Pkg.test("Gillespie"; coverage=true)'
