+++ docs/make.jl
@@ -5,7 +5,5 @@ makedocs(modules=[Gillespie],
 
 deploydocs(
     deps = Deps.pip("mkdocs","python-markdown-math"),
-    repo = "github.com/sdwfrost/Gillespie.jl.git",
-    julia = "0.4.6",
-    osname = "linux"
+    repo = "github.com/sdwfrost/Gillespie.jl.git"
 )
