+++ docs/mkdocs.yml
@@ -16,6 +16,7 @@ markdown_extensions:
   - extra
   - tables
   - fenced_code
+  - mdx_math
 
 docs_dir: 'build'
 
