### 2016-07-27 3ae4344
Updated docs
