### 2016-07-27 #42 @sdwfrost
I've added some documentation via Documenter.jl and updated the README to include proposed enhancements. I'm working on updating the benchmarks.
