+++ paper/paper.bib
@@ -9,7 +9,9 @@
   Number                   = {3},
   Pages                    = {455--473},
   Volume                   = {58},
-  Publisher                = {JSTOR}
+  Publisher                = {JSTOR},
+  doi                      = {10.2307/1990339},
+  url                      = {http://www.jstor.org/stable/1990339}
 }
 
 @Online{GillespieJL,
@@ -28,8 +30,9 @@
   Number                   = {25},
   Pages                    = {2340--2361},
   Volume                   = {81},
-  __markedentry            = {[simon:]},
   Publisher                = {ACS Publications}
+  doi                      = {10.1021/j100540a008},
+  url                      = {http://pubs.acs.org/doi/pdf/10.1021/j100540a008}
 }
 
 @Article{Pineda-Krch2008,
@@ -38,6 +41,7 @@
   Journal                  = {Journal of Statistical Software},
   Year                     = {2008},
   Pages                    = {12},
-  Volume                   = {25}
+  Volume                   = {25},
+  doi                      = {10.18637/jss.v025.i12},
+  url                      = {https://www.jstatsoft.org/article/view/v025i12}
 }
-
