+++ examples/decaying_dimer.jl
@@ -18,9 +18,9 @@ result = ssa(x0,F,nu,parms,tf)
 data = ssa_data(result)
 
 p=plot(data,
-  layer(x="time",y="x1",Geom.step,Theme(default_color=color("red"))),
-  layer(x="time",y="x2",Geom.step,Theme(default_color=color("blue"))),
-  layer(x="time",y="x3",Geom.step,Theme(default_color=color("green"))),
+  layer(x="time",y="x1",Geom.step,Theme(default_color=colorant"red")),
+  layer(x="time",y="x2",Geom.step,Theme(default_color=colorant"blue")),
+  layer(x="time",y="x3",Geom.step,Theme(default_color=colorant"green")),
   Guide.xlabel("Time"),
   Guide.ylabel("Number"),
   Guide.manual_color_key("Population",
