+++ src/Gillespie.jl
@@ -8,6 +8,7 @@ VERSION < v"0.4-dev" && using Lexicon
 export
 	ssa,
         ssa_data,
+        pfsample,
 	SSAArgs,
 	SSAStats,
 	SSAResult
