+++ docs/build/index.md
@@ -4,13 +4,214 @@
 # Gillespie.jl: Stochastic Simulation Algorithm in Julia
 
 - [Gillespie.jl: Stochastic Simulation Algorithm in Julia](index.md#Gillespie.jl:-Stochastic-Simulation-Algorithm-in-Julia-1)
-    - [Types](index.md#Types-1)
-    - [Functions](index.md#Functions-1)
+    - [Introduction](index.md#Introduction-1)
+    - [Installation](index.md#Installation-1)
+    - [Example](index.md#Example-1)
+    - [Application programming interface](index.md#Application-programming-interface-1)
+
+
+<a id='Introduction-1'></a>
+
+## Introduction
+
+
+`Gillespie.jl` provides an implementation of [Gillespie's direct method](http://en.wikipedia.org/wiki/Gillespie_algorithm) for performing stochastic simulations, which are widely used in many fields, including systems biology and epidemiology. It borrows the basic interface (although none of the code) from the R library [`GillespieSSA`](http://www.jstatsoft.org/v25/i12/paper) by Mario Pineda-Krch, although `Gillespie.jl` only implements the standard exact method at present, whereas `GillespieSSA` also includes tau-leaping, *etc.*.
+
+
+<a id='Installation-1'></a>
+
+## Installation
+
+
+`Gillespie.jl` can be installed from the Julia read-eval-print-loop (REPL) as follows.
+
+
+```julia
+Pkg.add("Gillespie")
+```
+
+
+<a id='Example-1'></a>
+
+## Example
+
+
+Let's take the 'standard' [susceptible-infected-recovered (SIR) model]((https://en.wikipedia.org/wiki/Compartmental_models_in_epidemiology#The_SIR_model_without_vital_dynamics)), commonly used in epidemiology. A deterministic version of this model is as follows.
+
+
+\begin{align} \frac{dS(t)}{dt}  & = -\beta S(t) I(t) \cr \frac{dI(t)}{dt}  & = \beta S(t) I(t)- \gamma I(t) \cr \frac{dR(t)}{dt}  & = \gamma I(t) \end{align}
+
+
+Let's consider a stochastic version of the SIR model.
+
+
+\begin{align} {\rm Transition} & \quad {\rm Rate} \cr S  \rightarrow S-1,\; I \rightarrow I+1 & \quad \beta S(t) I(t) \cr I  \rightarrow I-1,\; R \rightarrow R+1 & \quad \gamma I(t) \end{align}
+
+
+We first need to load the library.
+
+
+```julia
+using Gillespie;
+```
+
+
+We next need to define a function that given state variables `x` (type: `Array{Int64,1}`) and a vector of parameters (type: `Vector{Float64}`), returns a vector of rates of length `k` for different types of transitions. For this example, there are two transition functions, corresponding to infection and recovery.
+
+
+```julia
+function F(x,parms)
+  (S,I,R) = x
+  (beta,gamma) = parms
+  infection = beta*S*I
+  recovery = gamma*I
+  [infection,recovery]
+end;
+```
+
+```
+F (generic function with 1 method)
+```
+
+
+We define the states of the system - a `Vector{Int64}` of length `n` with the number of susceptible, infected, and recovered individuals.
+
+
+```julia
+x0 = [9999,1,0];
+```
+
+```
[…]
