### 2016-07-28 d0e05fa
Updated tests
