### 2016-07-28 05242e0
Fixing test suite
