### 2016-07-28 57b4f7a
Added benchmark file
