### 2016-07-28 #42 @sdwfrost
Dear @vchuravy @arfon @pjotrp

I think I've addressed all your comments now. Please let me know if there is anything I have missed. In summary

- Documentation is written using Documenter.jl
- Updated benchmarks are included for [Julia](https://gist.github.com/sdwfrost/8a0e926a5e16d7d104bd2bc1a5f9ed0b) and [R/Rcpp](https://gist.github.com/sdwfrost/afed3b881ef5742623b905a539197c7a). Although the plot doesn't show up in the IJulia notebook, the output is comparable between a handcoded version and `Gillespie.jl`. These benchmarks are summarised in the [README.md](https://github.com/sdwfrost/Gillespie.jl/blob/master/README.md). I hope that these results provide sufficient rationale for inclusion of `Gillespie.jl` in JOSS. I have also included a benchmark file in the repository.
- The tests now include checks that the output is at fixed for a given random number seed (in addition to checking that all the examples run). Code coverage is [95%](https://coveralls.io/github/sdwfrost/Gillespie.jl?branch=master)
- I've added some material to the end of the README with respect to future directions etc.
- dois added to the bib file.
- Release v0.0.1 added to the repository (rationale for numbering is following the Julia convention).
- Various badges have been added, including links to stable and latest docs, Coveralls.io, Waffle.io, and JOSS.

Thanks!
