### 2016-07-28 #42 @vchuravy
@arfon How would one regenerate the pdf? The .bib now includes DOI's, but they don't show up the the pdf.
@sdwfrost Great work! The codecoverage seems to be off again. One last question from my side. Have you tested compatibility with Julia `v0.5` (the rc0 was just announced)?

Otherwise this is a green light from me.
