### 2016-07-29 #42 @arfon
> @arfon How would one regenerate the pdf? The .bib now includes DOI's, but they don't show up the the pdf.

@vchuravy here's the updated PDF: [10.21105.joss.00042.pdf](https://github.com/openjournals/joss-reviews/files/389784/10.21105.joss.00042.pdf)

> @sdwfrost Great work! The codecoverage seems to be off again. One last question from my side. Have you tested compatibility with Julia v0.5 (the rc0 was just announced)?

@sdwfrost - once you've addressed this comment I think we're good to accept this submission 🎉
