### 2016-07-29 a769ab6
Update .travis.yml
Add nightly to Travis-CI
