+++ .travis.yml
@@ -4,8 +4,8 @@ os:
   - linux
 #  - osx
 julia:
-  - nightly
-  - 0.4
+  - 0.5.0-rc0
+  - 0.4.6
 notifications:
   email: false
 before_script:
