+++ .travis.yml
@@ -4,7 +4,7 @@ os:
   - linux
 #  - osx
 julia:
-  - 0.5.0-rc0
+  - 0.5
   - 0.4.6
 notifications:
   email: false
