### 2016-07-29 6880728
Fixed examples for v0.5
