+++ examples/sir2.jl
@@ -1,7 +1,6 @@
 using Gillespie
-using Gadfly
 
-function F(x,parms)
+function F_sir2(x,parms)
   (S,I,R) = x
   (beta,gamma) = parms
   infection = beta*S*I
@@ -14,20 +13,10 @@ nu = [[-1 1 0];[0 -1 1]]
 parms = [0.1/1000.0,0.01]
 tf = 250.0
 srand(1234)
-immutable G; end
-call(::Type{G},x,parms) = F(x,parms)
+immutable G_sir2; end
+call(::Type{G_sir2},x,parms) = F_sir2(x,parms)
 
-result = ssa(x0,G,nu,parms,tf)
+result = ssa(x0,G_sir2,nu,parms,tf)
 
 data = ssa_data(result)
 
-p=plot(data,
-  layer(x="time",y="x1",Geom.step,Theme(default_color=colorant"red")),
-  layer(x="time",y="x2",Geom.step,Theme(default_color=colorant"blue")),
-  layer(x="time",y="x3",Geom.step,Theme(default_color=colorant"green")),
-  Guide.xlabel("Time"),
-  Guide.ylabel("Number"),
-  Guide.manual_color_key("Population",
-                            ["S", "I", "R"],
-                            ["red", "blue", "green"]),
-  Guide.title("SIR epidemiological model"))
