### 2016-07-29 #42 @sdwfrost
@arfon @vchuravy Compatibility now checked against v0.4.6 and v0.5, all tests passing. I get some deprecation warnings on v0.5, but future releases will fix those.

@vchuravy I was getting some funny badges on Chrome for Coveralls (fine on Firefox) - click through, and you'll see coverage is still good.
