### 2016-07-30 #42 @arfon
@sdwfrost - as a final step, could you make an archive of the code in Zenodo/figshare? Please update this thread with the DOI of the code archive.
