+++ paper/paper.bib
@@ -19,7 +19,8 @@
   Author                   = {Simon D.W. Frost},
   URL                      = {https://github.com/sdwfrost/Gillespie.jl},
   Urldate                  = {2016-02-14},
-  Year                     = {2016}
+  Year                     = {2016},
+  doi                      = {http://dx.doi.org/10.5281/zenodo.59127}
 }
 
 @Article{Gillespie1977,
