### 2016-07-30 d6b3997
Update paper.bib
