### 2016-07-30 #42 @arfon
Thanks @sdwfrost. Your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00042 :rocket: 🎉 💥

Thanks for the review @vchuravy
